Implementation for all fields started. Infinite loop when searching for well-conditioned Y.

+++ KBessel.h
@@ -40,7 +40,7 @@ private:
 
     double r;
 
-    double numberOfShrinkingChunks;
+    int numberOfShrinkingChunks;
     double shrinkingChunkFirstWidth;
     double chunkWidth;
     double firstChunkLeftEndpoint;
